Fetch server version using futures
